requirements.txt
google-generativeai
python-dotenv
pypdf
numpy
